Keep repo minimal.
Now that I understand npm a little bit better I see why many authors
don't store `lib` and `dist` in the GitHub repo. These items are
generated on the fly using the `prepublish` hook and included in the npm
module. But they don't need to be stored in the GitHub repo.

I'm about 90% sure that's accurate at least.